+++ functions.php
@@ -2449,7 +2449,11 @@ function run_flow_query($session, $query_id, $start, $end) {
 
 				for ($i = 14; $i < 27; $i++) {
 					$name   = $sql_array[$i]['name'];
-					$table .= '<td class="right" style="width:7.14%">' . number_format_i18n(($results[$name] / $total) * 100, 2) . ' %</td>';
+					if ($total > 0) {
+						$table .= '<td class="right" style="width:7.14%">' . number_format_i18n(($results[$name] / $total) * 100, 2) . ' %</td>';
+					} else {
+						$table .= '<td class="right" style="width:7.14%">' . __('N/A', 'flowview') . '</td>';
+					}
 				}
 
 				$table .= '<td style="width:7.14%"></td>';